+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # v4
 - Bug fix: highlights are now exported in oldest to newest date order
 - Bug fix: Fixed "Last Highlight Date"
-- New: Date for the highlight is exported
+- New: Date heading for highlights
 
 # v3
 - Image  export